drop Exporter::Declare dependency
Exporter::Declare is rather weird and we have to do a lot of work
wrapping it anyway. It brings in a number of dependencies that are not
widely used. Replace it with an ad hoc solution.

This also allows being more consistent. Log::Contextual did not account
for all of the options Exporter::Declare provides, like renames.