pass importing package to subclass argument methods
The two included subclasses both want the importing package, but it was
not available directly. They were using caller to find this, but that
relies on the details of the import method implementation. Converting
away from Exporter::Declare changed the call stack, meaning the caller
level needed to be changed.

A much better option is to just pass the package along to the methods
for handling these arguments.